backend/requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-settings==2.8.1

# Database
sqlalchemy==2.0.40
sqlmodel==0.0.24
asyncpg==0.30.0

# API
python-multipart==0.0.20
httpx==0.28.1

# Agent Framework
agno==1.2.15
openai==1.72.0

# Supabase (Database and Auth)
supabase==2.15.0
postgrest==1.0.1

# Utilities
python-dateutil==2.9.0.post0
tqdm==4.67.1

# Development dependencies
black==25.1.0
flake8==7.2.0
pytest==8.3.5
pytest-mock==3.14.0
pre-commit==4.2.0
